+++ src/dwarfspec/config_schema.lua
@@ -15,6 +15,7 @@ local RESERVED_COMMANDS = {
     click=true,
     current_run=true,
     get=true,
+    getGameSpeed=true,
     getViewPos=true,
     hover=true,
     input=true,